Copy randomization
The old randomization logic would randomly select items from the array,
and then shrink it by shifting the elements down. This did not scale
well, since it spent a lot of time moving elements.

New logic just copies random elements and appends them to the end of a
new array. This is much faster than the previous version, but still not
the most efficient.

It's good enough for now.